requirements.txt
python-dotenv==1.0.0
supabase==1.0.3
aiohttp==3.8.5
beautifulsoup4==4.12.2
pandas==2.0.3
langchain-groq==0.0.3
langchain-mistralai==0.0.3
langchain-community==0.0.13
langchain==0.0.335
tenacity==8.2.3
ratelimit==2.2.1
tqdm==4.66.1
rich==13.5.2
lxml==4.9.3
mistralai==0.0.3
groq==0.4.1
asyncio==3.4.3
requests==2.31.0
python-dateutil==2.8.2
langchain-text-splitters==0.0.1 